--- lib/errors/index.spec.ts
@@ -1,9 +1,0 @@
-import * as allExports from '.';
-
-test('should have 5 exports', () => {
-    expect(Object.keys(allExports)).toHaveLength(5);
-});
-
-test('each of exports should be defined', () => {
-    Object.values(allExports).forEach(value => expect(value).toBeDefined());
-});
