+++ lib/redis/redis-manager.ts
@@ -3,7 +3,7 @@ import { Redis } from 'ioredis';
 import { RedisError } from 'redis-errors';
 import { REDIS_CLIENTS, DEFAULT_REDIS_NAMESPACE } from './redis.constants';
 import { RedisClients } from './interfaces';
-import { CLIENT_NOT_FOUND } from '@/errors';
+import { CLIENT_NOT_FOUND } from '@/messages';
 import { parseNamespace } from '@/utils';
 import { ClientNamespace } from '@/interfaces';
